archive/v1-microservices/doc-generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
structlog==23.2.0
weasyprint==60.2
python-docx==1.1.0
python-pptx==0.6.23
jinja2==3.1.2
httpx==0.25.2
pytest==7.4.3